--- src/main/java/br/com/cccat10/ecommerce/CreateOrderUseCase.java
@@ -1,7 +1,0 @@
-package br.com.cccat10.ecommerce;
-
-public interface CreateOrderUseCase {
-
-    void execute(Order order);
-
-}
