--- src/main/java/br/com/cccat10/ecommerce/ProductResponse.java
@@ -1,21 +1,0 @@
-package br.com.cccat10.ecommerce;
-
-import com.fasterxml.jackson.databind.PropertyNamingStrategies;
-import com.fasterxml.jackson.databind.annotation.JsonNaming;
-import lombok.Builder;
-import lombok.Data;
-
-import java.math.BigDecimal;
-
-@Data
-@Builder
-@JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class ProductResponse {
-
-    private String description;
-
-    private BigDecimal price;
-
-    private Integer quantity;
-
-}
