--- src/main/java/br/com/cccat10/ecommerce/RetrieveOrderUseCase.java
@@ -1,7 +1,0 @@
-package br.com.cccat10.ecommerce;
-
-public interface RetrieveOrderUseCase {
-
-    Order execute(Long id);
-
-}
