+++ src/main/java/br/com/cccat10/ecommerce/domain/response/ProductResponse.java
@@ -1,20 +1,21 @@
-package br.com.cccat10.ecommerce;
+package br.com.cccat10.ecommerce.domain.response;
 
 import com.fasterxml.jackson.databind.PropertyNamingStrategies;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
+import lombok.Builder;
 import lombok.Data;
 
 import java.math.BigDecimal;
-import java.util.List;
 
 @Data
+@Builder
 @JsonNaming(PropertyNamingStrategies.SnakeCaseStrategy.class)
-public class OrderRequest {
+public class ProductResponse {
 
-    private String buyerCpf;
+    private String description;
 
-    private List<ProductRequest> productList;
+    private BigDecimal price;
 
-    private BigDecimal discountPercentage;
+    private Integer quantity;
 
 }
