+++ src/main/java/br/com/cccat10/ecommerce/repository/OrderRepository.java
@@ -1,5 +1,6 @@
-package br.com.cccat10.ecommerce;
+package br.com.cccat10.ecommerce.repository;
 
+import br.com.cccat10.ecommerce.domain.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface OrderRepository extends JpaRepository<Order, Long> {
