+++ src/main/java/br/com/cccat10/ecommerce/usecase/CreateOrderUseCase.java
@@ -1,0 +1,9 @@
+package br.com.cccat10.ecommerce.usecase;
+
+import br.com.cccat10.ecommerce.domain.Order;
+
+public interface CreateOrderUseCase {
+
+    void execute(Order order);
+
+}
