+++ src/main/java/br/com/cccat10/ecommerce/usecase/RetrieveOrderUseCase.java
@@ -1,0 +1,9 @@
+package br.com.cccat10.ecommerce.usecase;
+
+import br.com.cccat10.ecommerce.domain.Order;
+
+public interface RetrieveOrderUseCase {
+
+    Order execute(Long id);
+
+}
