+++ src/main/java/br/com/cccat10/ecommerce/validator/CpfValidator.java
@@ -1,4 +1,4 @@
-package br.com.cccat10.ecommerce;
+package br.com.cccat10.ecommerce.validator;
 
 public interface CpfValidator {
